Refactor success page
