+++ app/javascript/components/donations/footer-form/FormModal.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useCallback } from 'react'
 import { Modal, ModalProps } from '../../modals/Modal'
 import currency from 'currency.js'
-import { DonationSuccess } from './stripe-form-modal/DonationSuccess'
+import SuccessModal from '../SuccessModal'
 import { Form } from '../Form'
 import { Request } from '../../../hooks/request-query'
 
@@ -23,6 +23,7 @@ export const FormModal = ({
   links: Links
 }): JSX.Element => {
   const [step, setStep] = useState<ModalStep>('donating')
+  const [paidAmount, setPaidAmount] = useState<currency | null>(null)
 
   const handleClose = useCallback(() => {
     if (step === 'processingDonation') {
@@ -36,7 +37,8 @@ export const FormModal = ({
     onClose()
   }, [onClose, step])
 
-  const handleDonationSuccess = useCallback(() => {
+  const handleDonationSuccess = useCallback((actualType, actualAmount) => {
+    setPaidAmount(actualAmount)
     setStep('donationSuccess')
   }, [])
 
@@ -52,26 +54,28 @@ export const FormModal = ({
   switch (step) {
     case 'donating':
     case 'processingDonation':
-      content = (
-        <Form
-          request={request}
-          defaultAmount={{ payment: amount, subscription: amount }}
-          defaultTransactionType="payment"
-          onSuccess={handleDonationSuccess}
-          links={links}
-          onProcessing={handleDonationProcessing}
-          onSettled={handleDonationSettled}
-        />
+      return (
+        <Modal className="m-donations-form" onClose={handleClose} {...props}>
+          <Form
+            request={request}
+            defaultAmount={{ payment: amount, subscription: amount }}
+            defaultTransactionType="payment"
+            onSuccess={handleDonationSuccess}
+            links={links}
+            onProcessing={handleDonationProcessing}
+            onSettled={handleDonationSettled}
+          />
+        </Modal>
       )
 
-      break
     case 'donationSuccess':
-      content = <DonationSuccess />
+      /* Kntsoriano - the amount here needs to be updated in the success callback */
+      return (
+        <SuccessModal
+          open={true}
+          amount={paidAmount}
+          closeLink={window.location.href}
+        />
+      )
   }
-
-  return (
-    <Modal className="m-donations-form" onClose={handleClose} {...props}>
-      {content}
-    </Modal>
-  )
 }
