--- app/javascript/components/donations/footer-form/stripe-form-modal/DonationSuccess.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export const DonationSuccess = (): JSX.Element => {
-  return <div>Thank you for donating!</div>
-}
